+++ ironfish/src/rpc/routes/wallet/utils.ts
@@ -5,8 +5,8 @@ import { Config } from '../../../fileStores'
 import { Note } from '../../../primitives'
 import { BufferUtils, CurrencyUtils } from '../../../utils'
 import { Account, Wallet } from '../../../wallet'
-import { isSignerMultiSig } from '../../../wallet/account/encoder/multisigKeys'
-import { MultisigKeys } from '../../../wallet/interfaces/multisigKeys'
+import { isSignerMultiSig } from '../../../wallet/account/encoder/multisigKey'
+import { MultisigKeys } from '../../../wallet/interfaces/multisigKey'
 import { AccountImport } from '../../../wallet/walletdb/accountValue'
 import { AssetValue } from '../../../wallet/walletdb/assetValue'
 import { DecryptedNoteValue } from '../../../wallet/walletdb/decryptedNoteValue'
