+++ ironfish/src/wallet/account/encoder/multiSigKeys.test.ts
@@ -1,8 +1,8 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { MultisigCoordinator, MultisigSigner } from '../../interfaces/multisigKeys'
-import { MultisigKeysEncoding } from './multisigKeys'
+import { MultisigCoordinator, MultisigSigner } from '../../interfaces/multisigKey'
+import { MultisigKeysEncoding } from './multisigKey'
 
 describe('multisigKeys encoder', () => {
   describe('with a signer value', () => {
