+++ ironfish/src/wallet/walletdb/accountValue.ts
@@ -5,8 +5,8 @@ import { PUBLIC_ADDRESS_LENGTH } from '@ironfish/rust-nodejs'
 import bufio from 'bufio'
 import { IDatabaseEncoding } from '../../storage'
 import { ACCOUNT_KEY_LENGTH } from '../account/account'
-import { MultisigKeysEncoding } from '../account/encoder/multisigKeys'
-import { MultisigKeys as MultisigKeys } from '../interfaces/multisigKeys'
+import { MultisigKeysEncoding } from '../account/encoder/multisigKey'
+import { MultisigKeys as MultisigKeys } from '../interfaces/multisigKey'
 import { HeadValue, NullableHeadValueEncoding } from './headValue'
 
 export const KEY_LENGTH = ACCOUNT_KEY_LENGTH
